Attach to session argument or list sessions
If given an argument, mx will treat it as a session and attempt to
attach to it. If the argument does not identify a session as per tmux's
standard session name matching, mx will instead list all current
sessions.

If more than one argument is given, all subsequent arguments are
ignored.

If no arguments are given, mx behaves as normal: attach to an existing
session identified by the working directory or create a new such session.