Ignore Vim's local spell-check dictionary
Several of these are sufficiently generic that distributing them might
be sensible, but I don't use spell-check enough to begin with that
figuring that out seems like a worthy time investment. Additionally,
several entries are not at all suitable for distribution, and really,
actually versioning the dictionary in the first place seems a little
questionable.