Do not auto-open vim-notes with mx
Auto-opening notes is extremely useful when the notes themselves are
needed. In practice, however, it has turned out that tmux is used at
least as much simply for the purpose of having simultaneous sessions
rather than emulating tabs in a terminal. Consequently, I ended up
favouring

    $ tmux new-session -As "${PWD##*/}"

to avoid the overhead of closing unsaved scratch files and numerous
windows caused by mx.

This patch removes the window that calls to vim when executing mx. Tig
is still opened. In the future, the ability to optionally open notes
and/or tig may prove worthwhile. In particular, mx no longer completely
solves the problem it was created for.