git-transplant for fast re-rebasing
git-transplant rebases all commits from HEAD up until the latest merge
commit onto the target ref.

It is a specialisation of

    git rebase --onto <ref1> <ref2>

where <ref1> is the target ref and <ref2> is the latest merge commit.
The argument to transplant is <ref1>.

Given

      B   D
     / \ /
    A---C

suppose B needs to be rewritten to B' and merged to create C':

      B   D
     / \ /
    A---C
     \
      B'---C'

B and C are now outdated but D needs to branch from C' instead:

      B'  D'
     / \ /
    A---C'

This will normally require rebase with the --onto flag and finding the
correct branching point. In this example, from branch D, the command

    git rebase --onto C' C

will suffice; but if there is no named pointer to C either an absolute
or a relative rev must be used. Instead, still from branch D, the
command

    git transplant C'

will suffice.