Use <Space> as <Leader>
<Space> is typically mapped to move-right, the same action as the l
motion key. I've used it as an alias for / for years, and that has
worked well because that action does not time out.

<Leader> does time out, so simply redefining g:mapleader as <Space> will
cause the original motion whenever the action times out. To avoid this,
<Space> must further be bound to <nop>. Additionally, either "<Space>"
must be escaped or a literal space character must be used:

    nnoremap <Space> <nop>
    " These are equivalent.
    let g:mapleader = " "
    let g:mapleader = "\<Space>"

Instead of this hassle, g:mapleader can be left alone and <Space> can be
aliased to it. This will not cause the normal motion on time out, and it
will display the command if showcmd is set.

Search has been rebound to <Leader><Space>, so double-space. This has
worked out okay so far, and the benefit from using <Space> as <Leader>
far outweighs the added difficulty.

Note that easymotion displayed some issues with this change. This added
to the decision to remove easymotion in "Remove vim-easymotion".