Enable YCM in plain-text files
YCM uses both a whitelist and a blacklist for supported file types, the
blacklist taking priority. The default setting disables YCM in, for
instance, Git commit messages and markdown files. For some reason simply
adjusting the list is not possible, so this commit takes the original
list, removes a number of plain-text file types for which completion is
desired, and uses that instead.

YCM is heavy, and a naive completion-centric solution would be preferred
over all the fast-feedback YCM offers, because Vim isn't being used for
the type of work that uses this anyway. neocomplete [1] does this, but
uses Lua and therefore is not compatible with Neovim. Shougu is working
on a Python 3 based implementation for Neovim [2], but it is not
progressing quickly.

[1] https://github.com/Shougo/neocomplete.vim
[2] https://github.com/Shougo/deoplete.nvim