Make vimrc Neovim compatible
Neovim supports 24 bit colours but it has to be enabled with an
environment variable. Unfortunately, the normal 256 colour Vim colouring
does not work at all when using Neovim inside tmux, which also only
supports 256 colours. This is all down to shitty colour support,
terminal emulators pretending to be terminals they are not, Neovim
replacing Vim's colouring, and it's a mess.

This commit makes Vim and Neovim colouring within and without tmux
correct, although Vim does not render quite as nicely as Neovim.