Fix E905 when setting encoding
Neovim made encoding only settable during startup and throws the new
E905 for attempts to set it later [1]. This vimrc auto-sources on save,
so every modification triggers an E905.

The change did not expose a way to detect this phase to prevent attempts
to set encoding during runtime; and encoding does need to be set,
because its default value depends on the environment. Possible solutions
are discussed in [2]; it turns out that vim_starting is a well-guarded
secret, mentioned only in :h initialization, that solves the problem
completely.

Latin-1 is still so widespread that the file encoding needs to be
allowed as a fall-back.

[1] https://github.com/neovim/neovim/pull/2929
[2] https://github.com/neovim/neovim/issues/3421