Always enable Neovim's 24 bit colour
Running Neovim with NVIM_TUI_ENABLE_TRUE_COLOR defined in vanilla tmux
<= 2.0 triggers this issue. Without that variable, Neovim will produce
poor but legible colouring. Vim works flawlessly either way.

This commit sets the relevant environment variable in vimrc
unconditionally, where before it was set only for xterm*. Although the
variable better belongs elsewhere, this option is most forgiving of my
work flow: an alias would not be picked up by third-party applications
(e.g. Git), necessitating further changes, and an external environment
variable would behave unpredictably in similar ways -- tig, for
instance, could not work with an adjusted $VISUAL.

The advantage of this approach is that Vim will ignore the setting and
Neovim will always enable true-colour support. The downside is that
Neovim will always enable true-colour support. That won't play nicely
with tmux core until they add support, too, but until then it can be
built from source.

Progress on adding true-colour support to tmux is being made [1], but
compiling that fork did not produce a binary that actually did support
true colours -- it behaved similarly to the unpatched version. Until
this work is finalised, a functional stand-alone patch and instructions
for applying it to tmux 2.0 are available instead [2].

A change in [3] conditionally set the environment variable in vimrc. At
the time, that produced as accurate colouring as possible; but it is
likely that an alias to Neovim, also setting that variable, was missed,
since on a fresh system that change alone turned out to be insufficient.

All of this applies to terminals without 24 bit colour support as well.

[1] https://github.com/tmux/tmux/pull/112
[2] https://gist.github.com/commonquail/643f64edbe58d5702844
[3] 658568e1644c72698db1f20f98a9a05be8cf6760