Revert "Define TERM"
This reverts commit 5e42c4f66afc5221d967985d45e445a92d05ae5e.

tmux dislikes any term that is not screen*, or in recent versions tmux*.
Careful reading of tmux's manual reveals this, but the situation is so
dire that to run tmux with any other value is considered undefined
behaviour -- the manual does not quite express this severity.

Starting tmux with no arguments starts a new shell. If that shell
explicitly defines TERM, and if that definition is anything other than
tmux's default-terminal setting -- which basically must be
screen-256color (or tmux-256color), the two settings will conflict but
tmux will not complain [1,2]. An example of where this causes breakage
is tig's stage view: line-wise staging occasionally but consistently
fails to refresh the screen below the cursor line.

To set TERM to anything other than screen-256color, it must be done in
the terminal emulator's own settings, and then tmux will still override
it. Explicitly setting TERM to screen-256color works but is kind of
pointless and unnecessary -- it will retard the regular terminal and
tmux will use the default-terminal setting anyway. Just pretend we're
not in the 21st century and leave the setting alone.

[1] http://askubuntu.com/a/426832
[2] http://superuser.com/a/399326/134756