Expose LESS_TERMCAP_* variables if colours are enabled
If colours are enabled, expose the LESS_TERMCAP_* variables so these are
always available to less. Order them sensibly and remove the redundant
man alias.

The primary effect of this change is make git help <command> behave as
man git <command>. A secondary effect is to not colour man pages in
colourless environments.