Avoid duplicate PATH entries
Define a function, pathprepend, that takes a single argument and adds it
to the front of PATH, iff that argument is a directory not already in
PATH. Based on http://superuser.com/a/39995/134756.

PATH is almost certainly marked for export by this point so it doesn't
need to be exported again.