Git: Add Reference command in gitcommit filetype
When editing commit messages, reference a commit with the command

    :Reference <rev>

in the style '<sha1> ("<subject>")' seen in the Linux kernel.

<rev> is a superset of the specifications detailed in gitrevisions(7).
The most useful of these is <rev>^{/<text>}, which yields the first
match of <text> in <rev>'s ancestors' commit messages, and the most
useful invocation of that is where <rev> is HEAD:

    Reference @^{/add fg}
    5000a98 ("Bash: do not add fg to history")

The far more convenient-to-type :/text form searches all refs and
consequently generates more false-positives. Therefore, as a special
case, if the argument to Reference begins with a slash (i.e. /text) --
in which case it is not a valid <rev> -- it is transformed into
HEAD^{/text}.