Vim: refactor commit reference command
The default-included patch output of "git show" can be suppressed,
making it a more appropriate choice than the workaround of limiting the
number of log entries. The official documentation includes an example
command to generate the reference output [1]. Although --format and
--pretty with placeholder strings should default to the tformat: variant
I didn't see any difference between the format: and tformat: variants
for this case.

[1] https://git.kernel.org/pub/scm/git/git.git/tree/Documentation/SubmittingPatches?h=v2.15.1#n139