Vim: work around broken jumplist
Neovim broke jumplist updates on "*" ("#") commands: if the cursor line
doesn't change the jumplist no longer updates. This came from commit
3536249 ("jumplist: avoid extra tail entry neovim#9805", 2019-04-02),
which indiscriminately deletes jumplist entries in some situations. It
was reported in neovim/neovim#9874, which was wontfix'ed on account of a
work-around of using "``" instead of "C-o". However, "``" is far from a
work-around: "``" also updates the jumplist so then unique matches move
the cursor instead of leaving it in place. neovim/neovim#12607 proposes
a trivial change that greatly improves the situation but still leaves
unique matches broken.

This commit records the issue and work-around for posterity but the
work-around does not work for me.