Git: correct commit reference format
Commit git@fb2ffa77a6 (SubmittingPatches: remove dq from commit
reference, 2019-11-19) pointed out that the format suggested up until
that point accounted for only 25% of occurrences and that the subject
line should not be quoted. Additionally, in Git v2.25.0 the "git log"
family learned the "--pretty=reference" format option as a short-hand
for precisely this usage [1]. openSUSE Leap 15.2 has v2.25.0 so the
current recommendation with that new format [2] is safe to use.

[1] https://git.kernel.org/pub/scm/git/git.git/tree/Documentation/RelNotes/2.25.0.txt?h=v2.25.0#n58
[2] https://git.kernel.org/pub/scm/git/git.git/tree/Documentation/SubmittingPatches?h=v2.25.0#n144